refactor(svelte-app): break app-state <-> manager service import cycle
cacheSecrets/cacheTimeout lived in app-state.ts, which the manager
service imported to push cache options into the SDK; app-state.ts in
turn imported the service for the storage handle. That cycle made
app-state's module-init touch the service before its `instance` was
initialized, risking a TDZ ReferenceError that was only papered over
with a settingsInitDone guard.

Extract the two stores into a leaf module (secret-cache-settings.ts,
depends only on svelte/store) imported by both sides. With the cycle
gone the TDZ cannot occur, so the settingsInitDone flag and its guard
are removed and resolveSettingsStorage() is simplified.
